requirements.txt
boto3==1.12.21
botocore==1.15.21
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
docutils==0.15.2
filelock==3.0.12
Flask==1.1.1
Flask-Cors==3.0.8
h5py==2.10.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.5
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
MarkupSafe==1.1.1
numpy==1.18.1
pandas==1.0.1
protobuf
python-dateutil
pytz==2019.3
PyYAML==5.3
regex==2020.2.20
requests==2.23.0
s3transfer==0.3.3
sacremoses==0.0.38
scikit-learn==0.22.2.post1
scipy==1.4.1
sentencepiece==0.1.85
seqeval==0.0.12
simpletransformers
six==1.14.0
tensorboardX==2.0
tokenizers
torch==1.4.0
tqdm
transformers
urllib3==1.25.8
Werkzeug==1.0.0
textract
python-docx